Code example in README.md would throw an exception
The Rollbar Person class (Person::setId() -> Utilities::validateString()) expects the id to be a string and will throw an exception if it is an integer. The original Context Information block contains code that would cause an exception to be thrown (line 92). This code has been modified so as to not throw an exception by casting the id parameter to a string.