Fix person_fn not being called when session not empty
Currently, when the session is not empty, this package always sets a
"person" entry in the rollbar config, which results in the rollbar/rollbar
package ever caling the person_fn.

The rollbar/rollbar package only calls $config['person_fn'] when the
$config['person'] doesn't already exist, which as stated above, will
always be set when the session is not empty.

Revert back to the behaviour originally found in RollbarLogHandler in
42920e5 before it was refactored in 8988c22 which introduced this bug.
As 42920e5 used an older version of the rollbar/rollbar library, this
commit doesn't revert the changes back exactly, but is the equivalent
changes for the latest version of the rollbar/rollbar library API.

Add a regression test that currently fails on master branch but passes
with these canges to RollbarLogHandler.